removing default quota override key, unifying test
